# Cache the escaped property names