--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterBoolean.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterBoolean.cs
@@ -25,9 +25,9 @@ namespace System.Text.Json.Serialization.Converters
             writer.WriteBooleanValue(value);
         }
 
-        public override void Write(Span<byte> escapedPropertyName, bool value, Utf8JsonWriter writer)
+        public override void Write(in JsonEncodedText propertyName, bool value, Utf8JsonWriter writer)
         {
-            writer.WriteBoolean(escapedPropertyName, value);
+            writer.WriteBoolean(propertyName, value);
         }
     }
 }