--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterBoolean.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterBoolean.cs
@@ -25,7 +25,7 @@ namespace System.Text.Json.Serialization.Converters
             writer.WriteBooleanValue(value);
         }
 
-        public override void Write(in JsonEncodedText propertyName, bool value, Utf8JsonWriter writer)
+        public override void Write(JsonEncodedText propertyName, bool value, Utf8JsonWriter writer)
         {
             writer.WriteBoolean(propertyName, value);
         }