--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterByte.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterByte.cs
@@ -27,9 +27,9 @@ namespace System.Text.Json.Serialization.Converters
             writer.WriteNumberValue(value);
         }
 
-        public override void Write(Span<byte> escapedPropertyName, byte value, Utf8JsonWriter writer)
+        public override void Write(in JsonEncodedText propertyName, byte value, Utf8JsonWriter writer)
         {
-            writer.WriteNumber(escapedPropertyName, value);
+            writer.WriteNumber(propertyName, value);
         }
     }
 }