--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterDateTimeOffset.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterDateTimeOffset.cs
@@ -24,7 +24,7 @@ namespace System.Text.Json.Serialization.Converters
             writer.WriteStringValue(value);
         }
 
-        public override void Write(in JsonEncodedText propertyName, DateTimeOffset value, Utf8JsonWriter writer)
+        public override void Write(JsonEncodedText propertyName, DateTimeOffset value, Utf8JsonWriter writer)
         {
             writer.WriteString(propertyName, value);
         }