--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterDecimal.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterDecimal.cs
@@ -24,7 +24,7 @@ namespace System.Text.Json.Serialization.Converters
             writer.WriteNumberValue(value);
         }
 
-        public override void Write(in JsonEncodedText propertyName, decimal value, Utf8JsonWriter writer)
+        public override void Write(JsonEncodedText propertyName, decimal value, Utf8JsonWriter writer)
         {
             writer.WriteNumber(propertyName, value);
         }