--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterEnum.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterEnum.cs
@@ -77,7 +77,7 @@ namespace System.Text.Json.Serialization.Converters
             }
         }
 
-        public override void Write(in JsonEncodedText propertyName, TValue value, Utf8JsonWriter writer)
+        public override void Write(JsonEncodedText propertyName, TValue value, Utf8JsonWriter writer)
         {
             if (TreatAsString)
             {