--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterSByte.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterSByte.cs
@@ -27,7 +27,7 @@ namespace System.Text.Json.Serialization.Converters
             writer.WriteNumberValue(value);
         }
 
-        public override void Write(in JsonEncodedText propertyName, sbyte value, Utf8JsonWriter writer)
+        public override void Write(JsonEncodedText propertyName, sbyte value, Utf8JsonWriter writer)
         {
             writer.WriteNumber(propertyName, value);
         }