--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterString.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterString.cs
@@ -25,9 +25,9 @@ namespace System.Text.Json.Serialization.Converters
             writer.WriteStringValue(value);
         }
 
-        public override void Write(Span<byte> escapedPropertyName, string value, Utf8JsonWriter writer)
+        public override void Write(in JsonEncodedText propertyName, string value, Utf8JsonWriter writer)
         {
-            writer.WriteString(escapedPropertyName, value);
+            writer.WriteString(propertyName, value);
         }
     }
 }