--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterString.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterString.cs
@@ -25,7 +25,7 @@ namespace System.Text.Json.Serialization.Converters
             writer.WriteStringValue(value);
         }
 
-        public override void Write(in JsonEncodedText propertyName, string value, Utf8JsonWriter writer)
+        public override void Write(JsonEncodedText propertyName, string value, Utf8JsonWriter writer)
         {
             writer.WriteString(propertyName, value);
         }