--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterUInt16.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterUInt16.cs
@@ -27,7 +27,7 @@ namespace System.Text.Json.Serialization.Converters
             writer.WriteNumberValue(value);
         }
 
-        public override void Write(in JsonEncodedText propertyName, ushort value, Utf8JsonWriter writer)
+        public override void Write(JsonEncodedText propertyName, ushort value, Utf8JsonWriter writer)
         {
             writer.WriteNumber(propertyName, value);
         }