--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterUInt32.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterUInt32.cs
@@ -24,7 +24,7 @@ namespace System.Text.Json.Serialization.Converters
             writer.WriteNumberValue(value);
         }
 
-        public override void Write(in JsonEncodedText propertyName, uint value, Utf8JsonWriter writer)
+        public override void Write(JsonEncodedText propertyName, uint value, Utf8JsonWriter writer)
         {
             writer.WriteNumber(propertyName, value);
         }