--- a/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterUInt64.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonValueConverterUInt64.cs
@@ -24,9 +24,9 @@ namespace System.Text.Json.Serialization.Converters
             writer.WriteNumberValue(value);
         }
 
-        public override void Write(Span<byte> escapedPropertyName, ulong value, Utf8JsonWriter writer)
+        public override void Write(in JsonEncodedText propertyName, ulong value, Utf8JsonWriter writer)
         {
-            writer.WriteNumber(escapedPropertyName, value);
+            writer.WriteNumber(propertyName, value);
         }
     }
 }