--- a/src/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfo.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfo.cs
@@ -38,7 +38,7 @@ namespace System.Text.Json.Serialization
         public string NameUsedToCompareAsString { get; private set; }
 
         // The escaped name passed to the writer.
-        public byte[] EscapedName { get; private set; }
+        public JsonEncodedText? EscapedName { get; private set; }
 
         public bool HasGetter { get; set; }
         public bool HasSetter { get; set; }
@@ -171,36 +171,7 @@ namespace System.Text.Json.Serialization
             }
 
             // Cache the escaped name.
-#if true
-            // temporary behavior until the writer can accept escaped string.
-            EscapedName = Name;
-#else
-            int valueIdx = JsonWriterHelper.NeedsEscaping(_name);
-            if (valueIdx == -1)
-            {
-                _escapedName = _name;
-            }
-            else
-            {
-                byte[] pooledName = null;
-                int length = JsonWriterHelper.GetMaxEscapedLength(_name.Length, valueIdx);
-
-                Span<byte> escapedName = length <= JsonConstants.StackallocThreshold ?
-                    stackalloc byte[length] :
-                    (pooledName = ArrayPool<byte>.Shared.Rent(length));
-
-                JsonWriterHelper.EscapeString(_name, escapedName, 0, out int written);
-
-                _escapedName = escapedName.Slice(0, written).ToArray();
-
-                if (pooledName != null)
-                {
-                    // We clear the array because it is "user data" (although a property name).
-                    new Span<byte>(pooledName, 0, written).Clear();
-                    ArrayPool<byte>.Shared.Return(pooledName);
-                }
-            }
-#endif
+            EscapedName = JsonEncodedText.Encode(Name);
         }
 
         private void DetermineSerializationCapabilities()