--- a/src/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfoNotNullable.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfoNotNullable.cs
@@ -78,18 +78,18 @@ namespace System.Text.Json.Serialization
 
             if (value == null)
             {
-                Debug.Assert(EscapedName != null);
+                Debug.Assert(EscapedName.HasValue);
 
                 if (!IgnoreNullValues)
                 {
-                    writer.WriteNull(EscapedName);
+                    writer.WriteNull(EscapedName.Value);
                 }
             }
             else if (ValueConverter != null)
             {
-                if (EscapedName != null)
+                if (EscapedName.HasValue)
                 {
-                    ValueConverter.Write(EscapedName, value, writer);
+                    ValueConverter.Write(EscapedName.Value, value, writer);
                 }
                 else
                 {