--- a/src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.HandleDictionary.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.HandleDictionary.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -124,37 +125,8 @@ namespace System.Text.Json.Serialization
             }
             else
             {
-                byte[] utf8Key = Encoding.UTF8.GetBytes(key);
-#if true
-                // temporary behavior until the writer can accept escaped string.
-                converter.Write(utf8Key, value, writer);
-#else
-                int valueIdx = JsonWriterHelper.NeedsEscaping(utf8Key);
-                if (valueIdx == -1)
-                {
-                    converter.Write(utf8Key, value, writer);
-                }
-                else
-                {
-                    byte[] pooledKey = null;
-                    int length = JsonWriterHelper.GetMaxEscapedLength(utf8Key.Length, valueIdx);
-
-                    Span<byte> escapedKey = length <= JsonConstants.StackallocThreshold ?
-                        stackalloc byte[length] :
-                        (pooledKey = ArrayPool<byte>.Shared.Rent(length));
-
-                    JsonWriterHelper.EscapeString(utf8Key, escapedKey, valueIdx, out int written);
-
-                    converter.Write(escapedKey.Slice(0, written), value, writer);
-
-                    if (pooledKey != null)
-                    {
-                        // We clear the array because it is "user data" (although a property name).
-                        new Span<byte>(pooledKey, 0, written).Clear();
-                        ArrayPool<byte>.Shared.Return(pooledKey);
-                    }
-                }
-#endif
+                JsonEncodedText escapedKey = JsonEncodedText.Encode(key);
+                converter.Write(escapedKey, value, writer);
             }
         }