Some clean-up of ifdef'd code and small perf gain (within margin of error) due to not having to check if the property name needs escaping.

Most of the converter changes will likely be removed in the future when\if we can set the property name independently.

Before:
```
|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 35.76 us | 0.8167 us | 0.9078 us | 35.64 us | 34.34 us | 37.96 us |      6.0069 |      0.4291 |           - |             25488 B |
| SerializeToUtf8Bytes | 33.20 us | 0.5414 us | 0.5065 us | 33.03 us | 32.34 us | 34.23 us |      3.0914 |           - |           - |             13016 B |
|    SerializeToStream | 32.17 us | 0.7758 us | 0.8301 us | 31.83 us | 31.40 us | 34.06 us |           - |           - |           - |               432 B |
```
After:
```
|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 34.79 us | 0.6817 us | 0.6377 us | 34.62 us | 34.07 us | 36.45 us |      6.0573 |           - |           - |             25568 B |
| SerializeToUtf8Bytes | 33.11 us | 0.4332 us | 0.3382 us | 33.15 us | 32.27 us | 33.61 us |      3.0371 |           - |           - |             12976 B |
|    SerializeToStream | 31.27 us | 0.4164 us | 0.3477 us | 31.13 us | 30.88 us | 32.13 us |           - |           - |           - |               432 B |
```