@ahsonkhan 2019-05-25
25488 vs 25568 - how come we end up allocating more for the string case?