@adamsitnik 2019-05-29
@AndyAyersMS @CarolEidt what is the easiest way to tell if `struct` is passed by ref/value in the generated machine code?