@CarolEidt 2019-05-29
> what is the easiest way to tell if struct is passed by ref/value in the generated machine code?

If you have a checked build and you dump the disassembly (`COMPlus_JitDisasm`) the call will have the primitive types of the arguments, e.g. `int`, `ref`, `byref`). If it is a struct passed by reference, it will be `byref`. However, if you're looking at the code in, say, a debugger, you'd have to reason about the code to figure out out.

That said, it will only list the argument as `byref` if it is specified as being passed by reference. If it is passed by making a copy and passing that by reference due to calling convention, then I don't think there's a straightforward way to tell.