@AndyAyersMS 2019-05-29
Also worth noting that the rules for what gets passed by value vs implicitly by reference depend on the ABI. So you can see different behavior on Linux vs Windows and on x86 vs x64 vs arm64. Things like argument position, struct size, and number and type of struct fields can factor into the decision.

For a smallish struct like `JsonEncodedText` (just two refs, right?) I would generally advocate passing it by value as it gives the jit the most room for optimization.