@ahsonkhan 2019-05-30
> 25488 vs 25568 - how come we end up allocating more for the string case?

@steveharter, can you share your benchmark code that you ran to get those results?