@adamsitnik 2019-05-30
From the method names, visible columns and reported time I guess that @steveharter has run these benchmarks from the performance repo: https://github.com/dotnet/performance/blob/master/src/benchmarks/micro/corefx/System.Text.Json/Serializer/WriteJson.cs

With the following filter

```cmd
--filter *WriteJson<IndexViewModel>*
```