@adamsitnik 2019-05-30
@steveharter I am currently profiling the JsonSerializer and trying to find ways to improve its perf. Escaping the property names twice was the first thing that popped out in the profiles, I have tried to fix it in #37977 but I was not aware of this PR. 
It would be much easier for me to continue the perf investigation if this PR is merged. All the reviewers have accepted it and the only remaining TODO is the "in" parameter.
To speed up the process I have cloned your fork, build it, removed the `in` and built again and compared. There is no difference when we use `in` so I have removed it and pushed the change to your fork. As soon as the CI gets green I am going to merge this PR which will simplify my perf investigation. I hope you don't mind.